Print Test renamed to PrintGYear Test GMonth.parse refactored (untested) GMonth.print implemented GMonth.reprint deleted (inherited)
