+++ nodejs/scripts/tests/GH73/GH73.js
@@ -4,7 +4,7 @@ module.exports = {
 		"GH73Core" : require('./GH73Core'),
 		"GH73DateObject" : require('./GH73DateObject'),
 		"GH73Functions" : require('./GH73Functions'),
-		"GH73Print" : require('./GH73Print'),
+		"GH73PrintGYear" : require('./GH73PrintGYear'),
 		"GH73Regex" : require('./GH73Regex')
 		
 	}
