+++ scripts/src/main/javascript/org/hisrc/jsonix/Jsonix/Schema/XSD/GYear.js
@@ -1,6 +1,8 @@
 Jsonix.Schema.XSD.GYear = Jsonix.Class(Jsonix.Schema.XSD.Calendar, {
 	name : 'GYear',
 	typeName : Jsonix.Schema.XSD.qname('gYear'),
+	CLASS_NAME : 'Jsonix.Schema.XSD.GYear',
+	
 	parse : function(value, context, input, scope) {
 		var gYearExpression = new RegExp("^" + Jsonix.Schema.XSD.Calendar.GYEAR_PATTERN + "$");
 		var results = value.match(gYearExpression);
@@ -11,11 +13,11 @@ Jsonix.Schema.XSD.GYear = Jsonix.Class(Jsonix.Schema.XSD.Calendar, {
 				date : this.xmlCalendarToDate(results[1], "01", "01", "00", "00", "00", results[5])
 			};
 			return splitedGYear;
-		} else {
-			throw new Error('Value [' + value + '] does not match the gYear pattern.');
 		}
+		throw new Error('Value [' + value + '] does not match the gYear pattern.');
 	},
-	print : function(value) {
+
+	print : function(value, context, input, scope) {
 		Jsonix.Util.Ensure.ensureObject(value);
 		var year = undefined;
 		var timezone = undefined;
@@ -28,15 +30,13 @@ Jsonix.Schema.XSD.GYear = Jsonix.Class(Jsonix.Schema.XSD.Calendar, {
 			year = value.year;
 			timezone = value.timezone;
 		}
-		// duschata quickfix to run GH73Print.js test
-		// must be placed in --> Calendar.printTimeZoneString()
-		// but has side effects in org.hisrc.jsonix.test.JsonixTest
+		// TODO: validation_issue (timezone range)
 		if (timezone < -14 * 60 || timezone > 14 * 60) {
 			throw new Error('Value ' + value + ' must not be <> -/+ ' + (14 * 60));
 		}
+		// TODO: validation_issue (signedYear)
 		return this.printSignedYear(year) + this.printTimeZoneString(timezone);
-	},
-	CLASS_NAME : 'Jsonix.Schema.XSD.GYear'
+	}
 });
 Jsonix.Schema.XSD.GYear.INSTANCE = new Jsonix.Schema.XSD.GYear();
 Jsonix.Schema.XSD.GYear.INSTANCE.LIST = new Jsonix.Schema.XSD.List(Jsonix.Schema.XSD.GYear.INSTANCE);
